requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
cx_Oracle==8.3.0
diff-match-patch==20241021
distro==1.9.0
Django==4.2.16
django-cors-headers==4.8.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==24.2
django-import-export==4.3.9
djangorestframework==3.16.1
drf-yasg==1.21.10
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.11.0
openai==1.107.2
oracledb==3.2.0
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
psutil==6.1.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyodbc==5.2.0
pypdfium2==4.30.0
python-decouple==3.8
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
sniffio==1.3.1
sqlparse==0.5.3
tablib==3.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.5.0


psycopg2-binary>=2.9.0